requirements.txt
gradio==3.44.1
numpy>=1.18.5,<1.24.0
cython
scipy
shapely
timm
h5py
wandb
fire
opencv-python
pandas
ftfy
regex
tqdm
gdown
Pillow==8.0
# Torch
#--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
#torch==1.10.1+cu113
#torchvision==0.11.2+cu113

# Detectron
#--find-links https://dl.fbaipublicfiles.com/detectron2/wheels/cu113/torch1.10/index.html
#detectron2

# Segment-anything
git+https://github.com/facebookresearch/segment-anything.git

# open_clip
open_clip_torch==1.3.0
